+++ src/index.ts
@@ -46,8 +46,8 @@ export * from './isNot';
 export * from './isNumber';
 export * from './isObject';
 export * from './isObjectType';
+export * from './isPlainObject';
 export * from './isPromise';
-export * from './isRecord';
 export * from './isString';
 export * from './isTruthy';
 export * from './join';
